feat: Player place connection

+++ DustRiders/DustRiders/Overlay.h
@@ -29,7 +29,7 @@ public:
 	void RenderWin(int windowHeight, int windowWidth);
 	void RenderLoss(int windowHeight, int windowWidth);
 	void RenderStorm(int frameWidth, int frameHeight, int screenWidth, int screenHeight);
-	void RenderSpeedometer(int currentSpeed, int screenWidth, int screenHeight);
+	void RenderSpeedometer(float currentSpeed, int screenWidth, int screenHeight);
 	void RenderCharge(int charge, int screenWidth, int screenHeight);
 	void RenderPlace(int place, int screenWidth, int screenHeight);
 	void Cleanup();
